Refactor ModuleSearchBox into reusable SearchBox
