+++ v3/src/js/views/components/SearchBox.scss
